% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_one.R
\name{add_one}
\alias{add_one}
\title{This function adds one to a number}
\usage{
add_one(number)
}
\arguments{
\item{number}{Numeric vector to which 1 will be added}
}
\value{
Numeric vector
}
\description{
This function adds one to a number
}
\examples{
add_one(5)
}
