% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two.R
\name{add_two}
\alias{add_two}
\title{This function adds two to a number}
\usage{
add_two(number)
}
\arguments{
\item{number}{Numeric vector to which 1 will be added}
}
\value{
Numeric vector
}
\description{
This function adds two to a number
}
\examples{
add_one(5)
}
